# Update relative perf margin in test_decoder_device_perf.py