Some perf targets were old and relative margins were too tight. This is failing TG model perf pipeline
https://github.com/tenstorrent/tt-metal/actions/runs/14515203700/job/40733975301

```
2025-04-17 16:21:06.340 | INFO     | models.demos.llama3_subdevices.tests.test_decoder_device_perf:test_llama_TG_perf_device:653 - RMSAllGather_3 op_to_op: 707.4444444444445 ns is larger than target (641.0) ns, difference: 2.3444444444444343 ns, margin: 0.1 ns, relative margin to pass would be: 0.10365747963251866
2025-04-17 16:21:06.341 | INFO     | models.demos.llama3_subdevices.tests.test_decoder_device_perf:test_llama_TG_perf_device:653 - AllGatherAsync_1 op_to_op: 959.5555555555555 ns is larger than target (796.5555555555555) ns, difference: 3.6888888888888687 ns, margin: 0.2 ns, relative margin to pass would be: 0.20463105035569815
2025-04-17 16:21:06.342 | INFO     | models.perf.benchmarking_utils:save_partial_run_json:177 - Run and measurement data saved to generated/benchmark_data/partial_run_2025-04-17T16:00:37+0000.json
Error: test_llama_TG_perf_device


2025-04-17 16:41:50.116 | INFO     | models.demos.llama3_subdevices.tests.test_decoder_device_perf:test_llama_TG_perf_device_non_overlapped_dispatch:802 - RMSAllGather_1 op_to_op: 6301.3 ns is smaller than target (7087.7) ns, difference: 77.63000000000011 ns, margin: 0.1 ns, relative margin to pass would be: 0.11095277734667094
2025-04-17 16:41:50.116 | INFO     | models.demos.llama3_subdevices.tests.test_decoder_device_perf:test_llama_TG_perf_device_non_overlapped_dispatch:792 - Matmul_3 op_to_op: 6879.1 ns is larger than target (6144.8) ns, difference: 119.81999999999971 ns, margin: 0.1 ns, relative margin to pass would be: 0.11949941413878404
2025-04-17 16:41:50.117 | INFO     | models.demos.llama3_subdevices.tests.test_decoder_device_perf:test_llama_TG_perf_device_non_overlapped_dispatch:792 - RotaryEmbeddingLlamaFusedQK_0 op_to_op: 3135.6 ns is larger than target (2844.3) ns, difference: 6.869999999999891 ns, margin: 0.1 ns, relative margin to pass would be: 0.10241535702984907
2025-04-17 16:41:50.117 | INFO     | models.demos.llama3_subdevices.tests.test_decoder_device_perf:test_llama_TG_perf_device_non_overlapped_dispatch:802 - PagedUpdateCacheDeviceOperation_0 op_to_op: 5006.4 ns is smaller than target (5890.0) ns, difference: 294.60000000000036 ns, margin: 0.1 ns, relative margin to pass would be: 0.15001697792869276
2025-04-17 16:41:50.118 | INFO     | models.perf.benchmarking_utils:save_partial_run_json:177 - Run and measurement data saved to generated/benchmark_data/partial_run_2025-04-17T16:21:17+0000.json
Error: test_llama_TG_perf_device_non_overlapped_dispatch
```

### Checklist
- [ ] [All post commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml) CI passes
- [ ] [Blackhole Post commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml) CI passes (if applicable)
- [ ] [Model regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-models.yaml) CI passes (if applicable)
- [ ] [Device performance regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-device-models.yaml) CI passes (if applicable)
- [ ] **(For models and ops writers)** Full [new models tests](https://github.com/tenstorrent/tt-metal/actions/workflows/full-new-models-suite.yaml) CI passes (if applicable)
- [ ] New/Existing tests provide coverage for changes